対応: [P1] 集計途中の整数精度喪失が相殺されると誤った収支を通常金額として表示する
自動開発サイクルで実装。AIレビュー、ローカルチェック、GitHub Actionsを通過。

+++ scripts/test-category.mjs
@@ -15,10 +15,10 @@ const result = context.categorySums([
 ]);
 
 const expected = new Map([
-  ["__proto__", 150],
-  ["constructor", 200],
-  ["日本語", 300],
-  ["未分類", 400],
+  ["__proto__", 150n],
+  ["constructor", 200n],
+  ["日本語", 300n],
+  ["未分類", 400n],
 ]);
 if (result.size !== expected.size || [...expected].some(([key, value]) => result.get(key) !== value)) {
   throw new Error("特殊なカテゴリ名の集計結果が想定と異なります");
@@ -32,8 +32,8 @@ const makeTransactions = count => Array.from({ length: count }, (_, index) => ({
 }));
 for (const count of [0, 8, 9]) {
   const display = context.categoryDisplayData(makeTransactions(count));
-  const displayedTotal = display.reduce((sum, [, value]) => sum + value, 0);
-  const expectedTotal = count * (count + 1) / 2;
+  const displayedTotal = display.reduce((sum, [, value]) => sum + value, 0n);
+  const expectedTotal = BigInt(count * (count + 1) / 2);
   if (displayedTotal !== expectedTotal || display.length !== (count === 9 ? 9 : count)) {
     throw new Error(`${count}カテゴリの表示合計が想定と異なります`);
   }
